requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
alembic==1.13.1
redis==5.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
httpx==0.27.0 